feat: add tenant management features and middleware
- Introduced CreateTenantDto and CreateUserDto for tenant and user creation.
- Implemented CloneTemplateSchemaUseCase for cloning database schemas.
- Added GetTenantByDomainUseCase to retrieve tenant information by domain.
- Created SlugifyNameUseCase for generating slugs from names.
- Developed middleware for tenant registration and subdomain handling.
- Implemented UserAgentService for parsing user agent strings.
- Added CommonController to expose user agent information endpoint.
- Created decorators for public access and role-based access control.
- Updated core module exports to include new functionalities.
