+++ apps/hr-core-server/src/modules/common/decorators/public.decorator.ts
@@ -1,4 +1,4 @@
-import { SetMetadata } from '@nestjs/common';
+// import { SetMetadata } from '@nestjs/common';
 
-export const IS_PUBLIC_KEY = 'isPublic';
-export const Public = () => SetMetadata(IS_PUBLIC_KEY, true);
+// export const IS_PUBLIC_KEY = 'isPublic';
+// export const Public = () => SetMetadata(IS_PUBLIC_KEY, true);
