+++ apps/hr-core-server/src/modules/common/decorators/roles.decorator.ts
@@ -1,4 +1,4 @@
-import { SetMetadata } from '@nestjs/common';
-import { RoleType } from '@kafaat-systems/entities';
+// import { SetMetadata } from '@nestjs/common';
+// import { RoleType } from '@kafaat-systems/entities';
 
-export const Roles = (...roles: RoleType[]) => SetMetadata('roles', roles);
+// export const Roles = (...roles: RoleType[]) => SetMetadata('roles', roles);
