+++ apps/hr-core-server/src/modules/common/midilwares/index.ts
@@ -1,24 +1,24 @@
-import { SubdomainMiddleware } from './subdomain.middleware';
-import { RequestMethod } from '@nestjs/common';
-import { TenantMiddleware } from './tenant.middlewar';
-import { RegisterTenantMiddleware } from './register-tenant.middlewar';
+// import { SubdomainMiddleware } from './subdomain.middleware';
+// import { RequestMethod } from '@nestjs/common';
+// import { TenantMiddleware } from './tenant.middlewar';
+// import { RegisterTenantMiddleware } from './register-tenant.middlewar';
 
-export const MIDDLEWARES = [
-  {
-    middleware: SubdomainMiddleware,
-    exclude: [
-      // { path: 'owner/(.*)', method: RequestMethod.ALL }
-    ],
-  },
-  {
-    middleware: TenantMiddleware,
-    exclude: [
-      { path: 'tenant/register', method: RequestMethod.POST },
-      { path: 'owner/(.*)', method: RequestMethod.ALL },
-    ],
-  },
-  {
-    middleware: RegisterTenantMiddleware,
-    include: [{ path: 'owner/register', method: RequestMethod.POST }],
-  },
-];
+// export const MIDDLEWARES = [
+//   {
+//     middleware: SubdomainMiddleware,
+//     exclude: [
+//       // { path: 'owner/(.*)', method: RequestMethod.ALL }
+//     ],
+//   },
+//   {
+//     middleware: TenantMiddleware,
+//     exclude: [
+//       { path: 'tenant/register', method: RequestMethod.POST },
+//       { path: 'owner/(.*)', method: RequestMethod.ALL },
+//     ],
+//   },
+//   {
+//     middleware: RegisterTenantMiddleware,
+//     include: [{ path: 'owner/register', method: RequestMethod.POST }],
+//   },
+// ];
