+++ apps/hr-core-server/src/modules/common/midilwares/subdomain.middleware.ts
@@ -1,132 +1,132 @@
-import { Injectable, NestMiddleware, Logger, HttpException, HttpStatus } from '@nestjs/common';
-import { Request, Response, NextFunction } from 'express';
-import { RoleType } from '@kafaat-systems/entities';
-import { SubdomainService } from '../services/subdomain.service';
-import { TenantContextService } from '@kafaat-systems/tenant-context';
-import { parse } from 'tldts';
-import { tenantContextStorage } from '@kafaat-systems/tenant-context';
-export interface SchemaRequest extends Request {
-  tenantId?: number;
-  schemaName?: string;
-  userRole?: RoleType;
-}
+// import { Injectable, NestMiddleware, Logger, HttpException, HttpStatus } from '@nestjs/common';
+// import { Request, Response, NextFunction } from 'express';
+// import { RoleType } from '@kafaat-systems/entities';
+// import { SubdomainService } from '../services/subdomain.service';
+// import { TenantContextService } from '@kafaat-systems/tenant-context';
+// import { parse } from 'tldts';
+// import { tenantContextStorage } from '@kafaat-systems/tenant-context';
+// export interface SchemaRequest extends Request {
+//   tenantId?: number;
+//   schemaName?: string;
+//   userRole?: RoleType;
+// }
 
-@Injectable()
-export class SubdomainMiddleware implements NestMiddleware {
-  private readonly logger = new Logger(SubdomainMiddleware.name);
+// @Injectable()
+// export class SubdomainMiddleware implements NestMiddleware {
+//   private readonly logger = new Logger(SubdomainMiddleware.name);
 
-  constructor(
-    private subdomainService: SubdomainService,
-    private tenantContextService: TenantContextService
-  ) {}
+//   constructor(
+//     private subdomainService: SubdomainService,
+//     private tenantContextService: TenantContextService
+//   ) {}
 
-  async use(req: SchemaRequest, res: Response, next: NextFunction): Promise<void> {
-    try {
-      // Check for role override in headers (for testing)
+//   async use(req: SchemaRequest, res: Response, next: NextFunction): Promise<void> {
+//     try {
+//       // Check for role override in headers (for testing)
 
-      if (process.env.NODE_ENV === 'development') {
-        const roleHeader = req.headers['x-user-role'] as string | undefined;
-        const schemaHeader = req.headers['x-schema-name'] as string | undefined;
+//       if (process.env.NODE_ENV === 'development') {
+//         const roleHeader = req.headers['x-user-role'] as string | undefined;
+//         const schemaHeader = req.headers['x-schema-name'] as string | undefined;
 
-        if (roleHeader) {
-          req.userRole = roleHeader as RoleType;
-          if (roleHeader === 'owner') {
-            this.logger.debug(`Using role from header: ${roleHeader}`);
-            next();
-            return;
-          }
+//         if (roleHeader) {
+//           req.userRole = roleHeader as RoleType;
+//           if (roleHeader === 'owner') {
+//             this.logger.debug(`Using role from header: ${roleHeader}`);
+//             next();
+//             return;
+//           }
 
-          this.logger.debug(`Using role from header1: ${roleHeader}`);
-        }
+//           this.logger.debug(`Using role from header1: ${roleHeader}`);
+//         }
 
-        if (schemaHeader) {
-          req.schemaName = schemaHeader;
-          this.logger.debug(`Using schema from header: ${schemaHeader}`);
+//         if (schemaHeader) {
+//           req.schemaName = schemaHeader;
+//           this.logger.debug(`Using schema from header: ${schemaHeader}`);
 
-          // Set the schema in the tenant context
-          if (req.schemaName) {
-            this.tenantContextService.setSchema(req.schemaName);
-          }
+//           // Set the schema in the tenant context
+//           if (req.schemaName) {
+//             this.tenantContextService.setSchema(req.schemaName);
+//           }
 
-          next();
-          return;
-        }
-      }
+//           next();
+//           return;
+//         }
+//       }
 
-      // Extract subdomain from host
-      const host = req.headers.host || '';
-      const parsedHost = parse(host);
+//       // Extract subdomain from host
+//       const host = req.headers.host || '';
+//       const parsedHost = parse(host);
 
-      // Remove localhost from host if present
-      const mainHost: string = '.' + process.env.BE_HOST || '';
-      let subdomain: string = parsedHost.hostname?.replace(mainHost, '') || '';
-      if (subdomain?.includes('www.') || subdomain?.includes('api.')) {
-        subdomain = subdomain?.replace('www.', '').replace('api.', '');
-      }
+//       // Remove localhost from host if present
+//       const mainHost: string = '.' + process.env.BE_HOST || '';
+//       let subdomain: string = parsedHost.hostname?.replace(mainHost, '') || '';
+//       if (subdomain?.includes('www.') || subdomain?.includes('api.')) {
+//         subdomain = subdomain?.replace('www.', '').replace('api.', '');
+//       }
 
-      // Check if we have a subdomain
-      if (subdomain) {
-        if (subdomain === 'owner') {
-          const context = {
-            schema: 'owner',
-            role: RoleType.OWNER,
-            tenantId: undefined,
-          };
+//       // Check if we have a subdomain
+//       if (subdomain) {
+//         if (subdomain === 'owner') {
+//           const context = {
+//             schema: 'owner',
+//             role: RoleType.OWNER,
+//             tenantId: undefined,
+//           };
 
-          tenantContextStorage.run(context, () => {
-            this.tenantContextService.setSchema('owner');
-            this.tenantContextService.setRole(RoleType.OWNER);
-            this.logger.debug('Using owner schema');
-            next();
-          });
+//           tenantContextStorage.run(context, () => {
+//             this.tenantContextService.setSchema('owner');
+//             this.tenantContextService.setRole(RoleType.OWNER);
+//             this.logger.debug('Using owner schema');
+//             next();
+//           });
 
-          return;
-        } else {
-          // Look up tenant by subdomain
+//           return;
+//         } else {
+//           // Look up tenant by subdomain
 
-          let tenant = await this.subdomainService.getTenantByDomain(subdomain);
-          // only login can check subdomain from email
-          if (!tenant && req.baseUrl.includes('login') && req.body.email) {
-            const email: string = req.body.email;
-            const domain = email.split('@')[1]?.split('.')[0];
-            tenant = await this.subdomainService.getTenantByDomain(domain);
-          }
-          if (tenant) {
-            if (!tenant?.isActive) {
-              throw new HttpException(
-                'Tenant is deactivated please contact support',
-                HttpStatus.BAD_REQUEST
-              );
-            }
-            const context = {
-              schema: tenant.schema_name,
-              role: RoleType.USER,
-              tenantId: String(tenant.id),
-            };
-            req.tenantId = parseInt(tenant.id);
-            req.schemaName = tenant.schema_name;
+//           let tenant = await this.subdomainService.getTenantByDomain(subdomain);
+//           // only login can check subdomain from email
+//           if (!tenant && req.baseUrl.includes('login') && req.body.email) {
+//             const email: string = req.body.email;
+//             const domain = email.split('@')[1]?.split('.')[0];
+//             tenant = await this.subdomainService.getTenantByDomain(domain);
+//           }
+//           if (tenant) {
+//             if (!tenant?.isActive) {
+//               throw new HttpException(
+//                 'Tenant is deactivated please contact support',
+//                 HttpStatus.BAD_REQUEST
+//               );
+//             }
+//             const context = {
+//               schema: tenant.schema_name,
+//               role: RoleType.USER,
+//               tenantId: String(tenant.id),
+//             };
+//             req.tenantId = parseInt(tenant.id);
+//             req.schemaName = tenant.schema_name;
 
-            tenantContextStorage.run(context, () => {
-              this.tenantContextService.setSchema(context.schema);
-              this.tenantContextService.setRole(context.role);
-              this.logger.debug(
-                `Tenant found for subdomain ${subdomain}, using schema: ${tenant.schema_name}`
-              );
-              next();
-            });
-          } else {
-            throw new HttpException('Tenant does not exist', HttpStatus.BAD_REQUEST);
-          }
-        }
-      } else {
-        throw new HttpException('Tenant does not exist', HttpStatus.BAD_REQUEST);
-      }
-    } catch (error) {
-      const message = error instanceof Error ? error.message : 'Unknown error';
+//             tenantContextStorage.run(context, () => {
+//               this.tenantContextService.setSchema(context.schema);
+//               this.tenantContextService.setRole(context.role);
+//               this.logger.debug(
+//                 `Tenant found for subdomain ${subdomain}, using schema: ${tenant.schema_name}`
+//               );
+//               next();
+//             });
+//           } else {
+//             throw new HttpException('Tenant does not exist', HttpStatus.BAD_REQUEST);
+//           }
+//         }
+//       } else {
+//         throw new HttpException('Tenant does not exist', HttpStatus.BAD_REQUEST);
+//       }
+//     } catch (error) {
+//       const message = error instanceof Error ? error.message : 'Unknown error';
 
-      this.logger.error(`Error in subdomain middleware: ${message}`);
-      // Continue with default schema
-      next(error);
-    }
-  }
-}
+//       this.logger.error(`Error in subdomain middleware: ${message}`);
+//       // Continue with default schema
+//       next(error);
+//     }
+//   }
+// }
