+++ apps/hr-core-server/src/modules/owner/owner.module.ts
@@ -1,15 +1,16 @@
-import { Module } from '@nestjs/common';
-import { OwnerController } from './owner.controller';
-import { OwnerService } from './owner.service';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { TenantEntity } from '@kafaat-systems/entities';
-import { CommonModule } from '../common/common.module';
-import { AuthModule } from '../auth/auth.module';
+// import { Module } from '@nestjs/common';
+// import { OwnerController } from './owner.controller';
+// import { OwnerService } from './owner.service';
+// import { TypeOrmModule } from '@nestjs/typeorm';
+// import { TenantEntity } from '@kafaat-systems/entities';
+// // import { CommonModule } from '../common/common.module';
+// // import { AuthModule } from '../auth/auth.module';
+// import { AuthModule, CommonModule } from '@kafaat-systems/core-app';
 
-@Module({
-  imports: [TypeOrmModule.forFeature([TenantEntity]), CommonModule, AuthModule],
-  controllers: [OwnerController],
-  providers: [OwnerService, CommonModule],
-  exports: [OwnerService],
-})
-export class OwnerModule {}
+// @Module({
+//   imports: [TypeOrmModule.forFeature([TenantEntity]), CommonModule, AuthModule],
+//   controllers: [OwnerController],
+//   providers: [OwnerService, CommonModule],
+//   exports: [OwnerService],
+// })
+// export class OwnerModule {}
