+++ libs/hr/core-app/src/lib/core/application/use-cases/slugify-name.use-case.ts
@@ -1,0 +1,8 @@
+import { Injectable } from '@nestjs/common';
+
+@Injectable()
+export class SlugifyNameUseCase {
+  async execute(name: string) {
+    return name.toLowerCase().replace(/\s+/g, '_');
+  }
+}
