+++ libs/hr/core-app/src/lib/core/interface/controllers/common.controller.ts
@@ -1,0 +1,12 @@
+import { Controller, Get, Headers } from '@nestjs/common';
+import { UserAgentService } from '../../infrastructure/services/user-agent.service';
+
+@Controller('common')
+export class CommonController {
+  constructor(private userAgentService: UserAgentService) {}
+
+  @Get('get-user-info')
+  getUserInfo(@Headers('user-agent') userAgent: string) {
+    return this.userAgentService.parse(userAgent);
+  }
+}
