+++ libs/hr/core-app/src/lib/core/interface/decorators/index.ts
@@ -1,0 +1,2 @@
+export * from './public.decorator';
+export * from './roles.decorator';
