+++ libs/hr/core-app/src/lib/core/interface/decorators/roles.decorator.ts
@@ -1,0 +1,4 @@
+import { SetMetadata } from '@nestjs/common';
+import { RoleType } from '@kafaat-systems/entities';
+
+export const Roles = (...roles: RoleType[]) => SetMetadata('roles', roles);
